HDUtilsTest: Move parse and format HDPath tests to HDPathTest.
