+++ core/src/test/java/org/bitcoinj/crypto/HDPathTest.java
@@ -16,10 +16,12 @@
 
 package org.bitcoinj.crypto;
 
+import com.google.common.collect.ImmutableList;
 import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.Collections;
+import java.util.List;
 
 /**
  * @author Michael Sean Gilligan
@@ -48,5 +50,61 @@ public class HDPathTest {
         Assert.assertEquals("m/0H/1H/0H",  path3.toString());
         Assert.assertEquals("m/0H/1H/0H/1",  path4.toString());
         Assert.assertEquals("m/0H/1H/0H/1/0",  path5.toString());
+    }
+
+    @Test
+    public void testFormatPath() {
+        Object[] tv = {
+                "M/44H/0H/0H/1/1",
+                ImmutableList.of(new ChildNumber(44, true), new ChildNumber(0, true), new ChildNumber(0, true),
+                        new ChildNumber(1, false), new ChildNumber(1, false)),
+
+                "M/7H/3/3/1H",
+                ImmutableList.of(new ChildNumber(7, true), new ChildNumber(3, false), new ChildNumber(3, false),
+                        new ChildNumber(1, true)),
+
+                "M/1H/2H/3H",
+                ImmutableList.of(new ChildNumber(1, true), new ChildNumber(2, true), new ChildNumber(3, true)),
+
+                "M/1/2/3",
+                ImmutableList.of(new ChildNumber(1, false), new ChildNumber(2, false), new ChildNumber(3, false))
+        };
+
+        for (int i = 0; i < tv.length; i += 2) {
+            String expectedStrPath = (String) tv[i];
+            HDPath path = HDPath.M((List<ChildNumber>) tv[i + 1]);
+
+            String generatedStrPath = path.toString();
+
+            Assert.assertEquals(generatedStrPath, expectedStrPath);
+        }
+    }
+
+    @Test
+    public void testParsePath() {
+        Object[] tv = {
+                "M / 44H / 0H / 0H / 1 / 1",
+                ImmutableList.of(new ChildNumber(44, true), new ChildNumber(0, true), new ChildNumber(0, true),
+                        new ChildNumber(1, false), new ChildNumber(1, false)),
+
+                "M/7H/3/3/1H/",
+                ImmutableList.of(new ChildNumber(7, true), new ChildNumber(3, false), new ChildNumber(3, false),
+                        new ChildNumber(1, true)),
+
+                "1 H / 2 H / 3 H /",
+                ImmutableList.of(new ChildNumber(1, true), new ChildNumber(2, true), new ChildNumber(3, true)),
+
+                "1 / 2 / 3 /",
+                ImmutableList.of(new ChildNumber(1, false), new ChildNumber(2, false), new ChildNumber(3, false))
+        };
+
+        for (int i = 0; i < tv.length; i += 2) {
+            String strPath = (String) tv[i];
+            List<ChildNumber> expectedPath = (List<ChildNumber>) tv[i + 1];
+
+            List<ChildNumber> path = HDPath.parsePath(strPath);
+
+            Assert.assertEquals(path, expectedPath);
+        }
     }
 }
